requirements.txt
discord.py[voice]==2.0.0
python-dotenv==1.0.0
yt-dlp
ytsearch
async-timeout==4.0.3
PyNaCl==1.5.0
aiohttp
ffmpeg-python
